requirements.txt
asgiref==3.7.2
dj-database-url==2.0.0
Django==4.2.2
django-ckeditor==6.6.1
django-js-asset==2.0.0
django-nested-admin==4.0.2
gunicorn==20.1.0
MarkupSafe==2.1.3
Pillow==9.5.0
psycopg2==2.9.6
python-monkey-business==1.0.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.7.0
tzdata==2023.3
whitenoise==6.5.0
